Fix bug: memory leaks in case of big messages (>2Gb). Fix bug: uncontrolled growth of the send pipeline buffer. Added max message size option.

+++ src/SharpRpc/Pipelines/RxPipeline.OneThread.cs
@@ -134,7 +134,7 @@ namespace SharpRpc
                     OnParseFailed(parseRes);       
             }
 
-            private void OnParseCompleted(int dataSize)
+            private void OnParseCompleted(long dataSize)
             {
                 TaskCompletionSource<bool> toSignal = null;
                 bool isNotClosed;
